umilib: Break out umilib.h
Break out umilib.h with a C-based subset of the umilib
interface. This is enough to created C-based wrappers around
umilib (e.g in QEMU) without code duplication.

No functional changes.
